2021 Edition (#62)
Starting to update the course materials. The 2021 edition formally introduces Colab for the first time, re-organizes some units, and slows the pace of the first few units.

There will be more updates over the next few months, and when all the material is stable again and tested, we can release the 2021 version.